Database support, new config file location
Links are now registered in a database !
Its configurtion muste be
defined in the database section in the config file. The migration for
the database can be done simply using "rake".

The config file has moved !
He must be located at "~/.playbot/config".

Site Plugins return value has changed !
The hash must contain a new key "url". It allow the plugin to normalize
it if many different url can link to a same content.